Add an Anki Question Formatter
This utility opens two successive vim-Instances, one for a question and
one for an answer.
You write the Question in the first, the answer in the second, each in
Markdown.
The utility (using pandoc and tr mostly) converts the markdown to html
and puts it all on one line, separating it by a semicolon and (if
specified) adding it to a file of questions.
This file can then ultimately be imported into Anki.