Start using neovim
Here the config is now under
  ~/.config/nvim/init.vim
This is currently the most up-to-date vim config that I have.
There is some necessary divergence from the .vimrc, most importantly
that the ugly cursor-shape-changer (abusing some silent `:!echo`
nonsense) can be yeeted because nvim does that by default. The upside
also being that now `:!!` actually works as it should.

nvim's real nice and I like it :)