qutebrowser: Add autoconfig-loading config
On this machine I ran into the error that having a 'config.py' file
(even if empty) would then not load the 'autoconfig.yml' at
qutebrowser startup.
I figure it's smarter to simply have the config.py file (which I
think got generated by a command like `:save` or something) with the
command to explicitly load the autoconfig, in which the mappings I
change using the commands in qutebrowser are automatically changed.