nvim: use pandoc's filetype 'markdown.pandoc'
Still annoyed with the pandoc-syntax linking its own highlight
groups e.g. because in Programming I'd like strings to be
deemphasized most of the time, but in markdown I'd like code
listings to be emphasized a bit (usually you look through an md-file
to find some command or something) and yet the plugin insists on
linking to String highlighting even if I've already set the plugins
highlight groups.