scripts::findpdf: Add a script to find a PDF file by content
I wrote this a while ago but apparently never pushed it, probably
because I felt it should just be an alias (but looking at it, theres
some complexity that merits a script here).
It uses `pdftotext` to get the PDF contents and then `fzf` to search
through them conveniently, ultimately providing us with the filename
of the result we selected. We still have to find the text in our PDF
viewer again, once we open the correct file.