scripts::explain_float: Add IEEE float explainer
Add a messy script to explain IEEE floats.
Very messy because I first wrote a quick script for single-precision
regular value floats and then felt I should add 16/64 bit support as
well as consider the special cases of zero, inf and NaN.
Bug rn stemming from the usage of awk as a calculator, but bc has
it's own issues as well, because it doesnt actually seem to use
floats (or something else weird is going on with the values).