bash: Set EDITOR to nvim properly
I needed this hack for stuff like meld, which would just call
editor, but this is neater. Maybe I'll opt for some bash magic to
figure out if we have a terminal and either call nvim directly or
open a terminal first in the case that its a GUI program calling.
For now, this is best.