scripts/theme-util: Try to be a bit more robust wrt. GTK theming
I noticed in Firefox the auto-theming wasn't working anymore.
I believe the reason was that at some point adwaita got uninstalled
and this script relied on using adwaita variants.
I found that the colorscheme thing was a good fallback, before
verifying that installing adwaita worked as well, so why not have
both, it's probably better.